requirements.txt
fastapi[all]==0.65.1
pyenchant==3.2.0
